Dashboard Cleanup
Cleaned up the dashboard so that there is no business logic in the razor page. Edited the queries in the controller to be slightly more readable. Created a new list containing competency trackers to make it easier to display information in the razor page. Edited razor page to only display the information relevant for client.